requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements.in
annotated-types==0.7.0
    # via pydantic
hyperlight-cocoa==0.2.1
    # via -r requirements.in
pydantic==2.10.6
    # via
    #   -r requirements.in
    #   hyperlight-cocoa
    #   taskex
pydantic-core==2.27.2
    # via pydantic
taskex==0.1.9
    # via -r requirements.in
typing-extensions==4.12.2
    # via
    #   pydantic
    #   pydantic-core
